test: split OCR failure tests into focused modules
